# api-docs.yml
...
swagger: '2.0'
info:
  description: 칵테일 공유 커뮤니티 마쉴랭의 API 명세서입니다.
  title: Maswillaeng
...
paths:
...
  "/api/change-format-image":
    post:
      tags:
        - post-controller
      summary: changeImageFormat
      operationId: changeImageFormatUsingPOST
      consumes:
        - multipart/form-data
      produces:
        - "*/*"
      parameters:
        - in: body
          name: photo
          description: photo
          required: true
          schema:
            type: string
            format: binary
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
...
  "/api/post":
    get:
      tags:
        - post-controller
      summary: getPostList
      operationId: getPostListUsingGET
      produces:
        - "*/*"
      parameters:
        - name: category
          in: query
          description: category
          required: false
          type: string
          enum:
            - BAR_SNACK
            - BEST
            - FREE
            - RECIPE
      responses:
        '200':
          description: OK
          schema:
            type: object
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    post:
      tags:
        - post-controller
      summary: savePost
      operationId: savePostUsingPOST
      consumes:
        - application/json
      produces:
        - "*/*"
      parameters:
        - in: body
          name: requestDto
          description: requestDto
          required: true
          schema:
            "$ref": "#/definitions/PostRequestDto"
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    put:
      tags:
        - post-controller
      summary: updatePost
      operationId: updatePostUsingPUT
      consumes:
        - application/json
      produces:
        - "*/*"
      parameters:
        - in: body
          name: updateDto
          description: updateDto
          required: true
          schema:
            "$ref": "#/definitions/PostUpdateDto"
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
...
  "/api/post/report":
    get:
      tags:
        - post-controller
      summary: getReportedPostList
      operationId: getReportedPostListUsingGET
      produces:
        - "*/*"
      parameters:
        - name: page
          in: query
          description: page
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          schema:
            type: object
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  "/api/post/user":
    get:
      tags:
        - post-controller
      summary: getUserPostList
      operationId: getUserPostListUsingGET
      produces:
        - "*/*"
      parameters:
        - name: category
          in: query
          description: category
          required: false
          type: string
        - name: page
          in: query
          description: page
          required: true
          type: integer
          format: int32
        - name: userId
          in: query
          description: userId
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            type: object
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  "/api/post/{postId}":
    get:
      tags:
        - post-controller
      summary: getPost
      operationId: getPostUsingGET
      produces:
        - "*/*"
      parameters:
        - name: postId
          in: path
          description: postId
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            type: object
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
        - post-controller
      summary: deletePost
      operationId: deletePostUsingDELETE
      produces:
        - "*/*"
      parameters:
        - name: postId
          in: path
          description: postId
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            type: object
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
  "/api/user":
    get:
      tags:
        - user-controller
      summary: getUserApi
      operationId: getUserApiUsingGET
      produces:
        - "*/*"
      responses:
        '200':
          description: OK
          schema:
            type: object
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    put:
      tags:
        - user-controller
      summary: updateUserInfo
      operationId: updateUserInfoUsingPUT
      consumes:
        - application/json
      produces:
        - "*/*"
      parameters:
        - in: body
          name: requestDto
          description: requestDto
          required: true
          schema:
            "$ref": "#/definitions/UserUpdateRequestDto"
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
        - user-controller
      summary: userWithDraw
      operationId: userWithDrawUsingDELETE
      produces:
        - "*/*"
      responses:
        '200':
          description: OK
          schema:
            type: object
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  "/api/user/image":
    post:
      tags:
        - user-controller
      summary: userImageUpdate
      operationId: userImageUpdateUsingPOST
      consumes:
        - multipart/form-data
      produces:
        - "*/*"
      parameters:
        - in: body
          name: photo
          description: photo
          required: true
          schema:
            type: string
            format: binary
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  "/api/user/reset-pwd":
    put:
      tags:
        - user-controller
      summary: resetPassword
      operationId: resetPasswordUsingPUT
      consumes:
        - application/json
      produces:
        - "*/*"
      parameters:
        - in: body
          name: requestDto
          description: requestDto
          required: true
          schema:
            "$ref": "#/definitions/UserPwdResetRequestDto"
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  "/api/userInfo/{userId}":
    get:
      tags:
        - user-controller
      summary: getUserInfo
      operationId: getUserInfoUsingGET
      produces:
        - "*/*"
      parameters:
        - name: userId
          in: path
          description: userId
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            type: object
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
...
definitions:
...
  PostRequestDto:
    type: object
    required:
      - content
      - title
    properties:
      category:
        type: string
      content:
        type: string
      hashTagList:
        type: array
        items:
          type: string
      thumbnail:
        type: string
      title:
        type: string
    title: PostRequestDto
  PostUpdateDto:
    type: object
    required:
      - content
      - title
    properties:
      category:
        type: string
        enum:
          - BAR_SNACK
          - BEST
          - FREE
          - RECIPE
      content:
        type: string
      hashTagList:
        type: array
        items:
          type: string
      postId:
        type: integer
        format: int64
      thumbnail:
        type: string
      title:
        type: string
    title: PostUpdateDto
...
  UserPwdResetRequestDto:
    type: object
    properties:
      email:
        type: string
      password:
        type: string
      token:
        type: string
    title: UserPwdResetRequestDto
  UserUpdateRequestDto:
    type: object
    properties:
      introduction:
        type: string
      nickName:
        type: string
      userImage:
        type: string
    title: UserUpdateRequestDto
...